+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/session/DefaultMeetingSession.kt
@@ -48,6 +48,7 @@ class DefaultMeetingSession(
 
         val audioClientController =
             DefaultAudioClientController(
+                context,
                 logger,
                 audioClientObserver,
                 audioClient